more cleanup and crap
Cleaned up some more code, organized which heists disable the mod better.

Added Diamond Heist to the list of heists that are always night.

Disabled the mod on Safehouse Nightmare and Lab Rats as it did nothing but take up memory.

Disabled some incorrect environment replacements for heists the mod is disabled on.

Fixed an environment issue with undercover.