change outfits, disable fwb
Added a new option to force outfits to match the weather, this is done clientside.
Disabled First World Bank from being affected by the mod. It had a ton of problems and you're only outside 5 seconds anyway.
Fixed snow using the same random roll as rain.
If the snow option is on, the mod will now force snow on boiling point.